upload avatar, follow, like

+++ app/src/main/java/com/example/tiktok/adapters/VideoGridAdapter.java
@@ -1,6 +1,9 @@
 package com.example.tiktok.adapters;
 
+import static androidx.core.content.ContentProviderCompat.requireContext;
+
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
 import android.util.Log;
 import android.view.View;
@@ -10,21 +13,30 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.example.tiktok.MainActivity;
 import com.example.tiktok.R;
+import com.example.tiktok.models.Root;
 import com.example.tiktok.models.Video;
+import com.example.tiktok.service.ApiInterface;
+import com.example.tiktok.service.RetrofitClient;
 import com.example.tiktok.utils.MyUtil;
 
+import java.util.Iterator;
 import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 public class VideoGridAdapter extends RecyclerView.Adapter<VideoGridAdapter.ViewHolder> {
     private static final String TAG = "VideoGridAdapter";
 
     List<Video> videos;
     Context context;
+    final ApiInterface apitiktok = RetrofitClient.getInstance().create(ApiInterface.class);
     public VideoGridAdapter(List<Video> videos, Context context) {
         this.videos = videos;
         this.context = context;
@@ -41,21 +53,24 @@ public class VideoGridAdapter extends RecyclerView.Adapter<VideoGridAdapter.View
 
         updateView(holder, video);
 
-        //hide delete button if user is not author
-//        if (MainActivity.getCurrentUser().getUsername().equals(video.getUsername()) || MainActivity.getCurrentUser().isAdmin()) {
+        //hide delete button if user xem
+        if (MainActivity.getCurrentUser().getUser_id().equals(video.getUser_id())) {
             holder.txt_delete_video.setOnClickListener(v -> handleDeleteVideo(video, position));
-//        } else {
-//            holder.txt_delete_video.setVisibility(View.GONE);
-//        }
+        } else {
+            holder.txt_delete_video.setVisibility(View.GONE);
+        }
+
+        holder.txt_delete_video.setOnClickListener(v -> handleDeleteVideo(video, position));
     }
 
     private void updateView(ViewHolder holder, Video video) {
         Log.i(TAG, "updateView: " + video.getContent());
         holder.txt_num_likes.setText(String.valueOf(video.getNum_like()));
         holder.txt_num_views.setText(String.valueOf(video.getNum_views()));
+        String videoUrl = RetrofitClient.getBaseUrl() +"/api/file/video/view?fileName=" + video.getFileName();
         try {
             Glide.with(context)
-                    .load(video.getFileName())
+                    .load(videoUrl)
                     .error(R.drawable.img_404)
                     .into(holder.img_preview);
         } catch (Exception e) {
@@ -74,10 +89,21 @@ public class VideoGridAdapter extends RecyclerView.Adapter<VideoGridAdapter.View
         builder.setTitle("Xóa video")
                 .setMessage("Bạn có chắc chắn muốn xóa video này?")
                 .setPositiveButton("Có", (dialog, which) -> {
-//                    VideoFirebase.deleteVideo(video);
-                    Toast.makeText(context, "Xóa video thành công", Toast.LENGTH_SHORT).show();
-                    videos.remove(position);
-                    notifyItemRemoved(position);
+                    apitiktok.deletevideo(video.getVideo_id()).enqueue(new Callback<Root<Video>>() {
+                        @Override
+                        public void onResponse(Call<Root<Video>> call, Response<Root<Video>> response) {
+                            if(response.isSuccessful()){
+                                Toast.makeText(context, "Xóa video thành công", Toast.LENGTH_SHORT).show();
+                                videos.remove(position);
+                                notifyItemRemoved(position);
+                            }
+                        }
+
+                        @Override
+                        public void onFailure(Call<Root<Video>> call, Throwable t) {
+                            Log.e("Xóa thất bại", t.getMessage());
+                        }
+                    });
                 })
                 .setNegativeButton("Hủy", (dialog, which) -> dialog.dismiss())
                 .show();
@@ -100,6 +126,7 @@ public class VideoGridAdapter extends RecyclerView.Adapter<VideoGridAdapter.View
         return videos.size();
     }
     public void setVideos(List<Video> newVideos) {
+        //kiểm tra video đã có trong danh sách chưa
         for (Video video : newVideos) {
             if (!videos.contains(video)) {
                 videos.add(video);
@@ -109,11 +136,22 @@ public class VideoGridAdapter extends RecyclerView.Adapter<VideoGridAdapter.View
                 notifyItemChanged(index, video);
             }
         }
-
-        for (Video video : videos) {
+//        videos.removeAll(newVideos);
+//        notifyDataSetChanged();
+//        for (Video video : videos) {
+//            if (!newVideos.contains(video)) {
+//                int index = videos.indexOf(video);
+//                videos.remove(video);
+//                notifyItemChanged(index, video);
+////                break;
+//            }
+//        }
+        Iterator<Video> iterator = videos.iterator();
+        while (iterator.hasNext()) {
+            Video video = iterator.next();
             if (!newVideos.contains(video)) {
                 int index = videos.indexOf(video);
-                videos.remove(video);
+                iterator.remove(); // Sử dụng iterator để xóa video an toàn
                 notifyItemRemoved(index);
             }
         }
